pyproject.toml
[project]
name = "mini-modelvault"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.5",
    "loguru>=0.7.3",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "rich>=14.0.0",
    "typer>=0.16.0",
    "uvicorn>=0.35.0",
    "yaspin>=3.1.0",
]

[project.scripts]
mini_modelvault = "mini_modelvault.main:main"

[tool.setuptools]
packages = { find = {} }

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"